+++ src/main/java/com/webapi/contacts/service/SkillService.java
@@ -1,8 +1,13 @@
 package com.webapi.contacts.service;
 
+import com.webapi.contacts.exception.UnchangableSkillException;
 import com.webapi.contacts.model.Skill;
+import com.webapi.contacts.model.User;
 import com.webapi.contacts.repository.SkillRepository;
+import com.webapi.contacts.service.auth.CustomUserPrincipal;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import javax.persistence.EntityNotFoundException;
@@ -31,12 +36,31 @@ public class SkillService {
     }
 
     public Skill updateSkill(Skill skillToUpdate) {
-        skillRepository.findById(skillToUpdate.getSkillId()).orElseThrow(EntityNotFoundException::new);
-        return skillRepository.save(skillToUpdate);
+        Skill existingSkill = skillRepository.findById(skillToUpdate.getSkillId()).orElseThrow(EntityNotFoundException::new);
+        User userFromContactToUpdate = existingSkill.getContacts().get(0).getUser();
+        User userFromContext = getUserFromContext();
+
+        if (userFromContext.equals(userFromContactToUpdate)) {
+            return skillRepository.save(skillToUpdate);
+        } else {
+            throw new UnchangableSkillException();
+        }
     }
 
     public void deleteSkillForId(Long id) {
         Skill existingSkill = skillRepository.findById(id).orElseThrow(EntityNotFoundException::new);
-        skillRepository.delete(existingSkill);
+        User userFromContactToUpdate = existingSkill.getContacts().get(0).getUser();
+        User userFromContext = getUserFromContext();
+
+        if (userFromContext.equals(userFromContactToUpdate)) {
+            skillRepository.delete(existingSkill);
+        } else {
+            throw new UnchangableSkillException();
+        }
+    }
+
+    private User getUserFromContext() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        return ((CustomUserPrincipal) authentication.getPrincipal()).getUser();
     }
 }
